+++ client/src/components/CartPage/CartPage.jsx
@@ -1,4 +1,7 @@
 import React from 'react';
+
+import styles from './CartPage.module.scss';
+
 import ShoppingBag from 'components/ShoppingBag';
 import Button from 'components/Button';
 
@@ -6,7 +9,7 @@ function CartPage() {
     return (
         <>
             <ShoppingBag />
-            <Button text="PROCEED TO CHECKOUT" />
+            <Button text="PROCEED TO CHECKOUT" className={styles.checkoutBtn} />
         </>
     );
 }
