+++ client/src/components/CartPage/CartPage.module.scss
@@ -1,0 +1,14 @@
+@import 'index.scss';
+
+.checkoutBtn {
+    @include abel-font-normal;
+
+    display: block;
+    font-size: 1.8rem;
+    line-height: 2rem;
+    padding: 1.5rem 3.6rem;
+    color: rgb(255, 255, 255);
+    background-color: $dark-color;
+    border-radius: 4px;
+    margin: 5rem auto 0 ;
+}
