+++ client/src/components/ShoppingBag/ShoppingBag.module.scss
@@ -1,4 +1,5 @@
 @import 'style/_mixins.scss';
+@import './style';
 
 .bagWrapper {
     @include content-container;
@@ -10,6 +11,19 @@
     justify-content: space-between;
     width: 100%;
     margin-bottom: 20px;
+
+    .bagHeader {
+        @include muli-font-normal;
+        color: $mainTxtColor;
+        font-size: 2.4rem;
+        line-height: 2rem;
+    }
+    .totalPrice {
+        @include muli-font-semibold;
+        color: $mainTxtColor;
+        font-size: 2.4rem;
+        line-height: 2rem;
+    }
 }
 
 .itemsList {
