+++ client/src/components/ShoppingBag/ShoppingBag.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import ShoppingBag from './ShoppingBag';
+
+test('button renders with props', () => {
+    render(<ShoppingBag />);
+    // const linkElement = screen.
+    // expect(linkElement).toBeInTheDocument();
+});
