--- client/src/components/ShoppingBag/components/BagItemForm/index.js
@@ -1,1 +1,0 @@
-export { default } from './BagItemForm';
