--- client/src/components/ShoppingBag/components/Header/Header.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './Header.module.scss';
-
-const Header = () => {
-    return <h2 className={styles.bagHeader}>SHOPPING BAG</h2>;
-};
-
-export default Header;
