--- client/src/components/ShoppingBag/components/Header/Header.module.scss
@@ -1,10 +1,0 @@
-@import 'style/_mixins.scss';
-@import '../../style';
-
-.bagHeader {
-    @include muli-font-normal;
-    color: $mainTxtColor;
-    font-size: 2.4rem;
-    line-height: 2rem;
-}
-
