+++ client/src/components/ShoppingBag/components/QuantityInput/index.js
@@ -1,0 +1,1 @@
+export { default } from './QuantityInput';
