--- client/src/components/ShoppingBag/components/TotalPrice/index.js
@@ -1,1 +1,0 @@
-export { default } from './TotalPrice';
